Add BpkCarouselInternalState
